Divide up ribbon and wxmribbon objects for gui specific and gui abstract
